third_party/blink/web_tests/platform/generic/http/tests/eventsource/workers/eventsource-cors-basic-expected.txt

CONSOLE ERROR: Access to resource at 'http://127.0.0.1:8080/eventsource/resources/es-cors-basic.php?count=1' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: Access to resource at 'http://127.0.0.1:8080/eventsource/resources/es-cors-basic.php?count=2' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: The 'Access-Control-Allow-Origin' header has a value 'http://some.other.origin:80' that is not equal to the supplied origin.
[Worker] Test that basic EventSource cross-origin requests fail until they are allowed by the Access-Control-Allow-Origin header.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: ../script-tests/eventsource-cors-basic.js
PASS [Worker] es = new EventSource("http://127.0.0.1:8080/eventsource/resources/es-cors-basic.php?count=1") did not throw exception.
PASS [Worker] es.withCredentials is false
PASS [Worker] es.readyState is es.CLOSED
PASS [Worker] count != 3 && count != 4 is true
PASS [Worker] es = new EventSource("http://127.0.0.1:8080/eventsource/resources/es-cors-basic.php?count=2") did not throw exception.
PASS [Worker] es.withCredentials is false
PASS [Worker] es.readyState is es.CLOSED
PASS [Worker] count != 3 && count != 4 is true
PASS [Worker] es = new EventSource("http://127.0.0.1:8080/eventsource/resources/es-cors-basic.php?count=3") did not throw exception.
PASS [Worker] es.withCredentials is false
PASS [Worker] evt.origin.indexOf('http://127.0.0.1:8080') === 0 is true
PASS [Worker] evt.data == 'DATA1' && evt.lastEventId == '77' is true
PASS [Worker] count is 4
PASS [Worker] evt.origin.indexOf('http://127.0.0.1:8080') === 0 is true
PASS [Worker] count is 4
PASS [Worker] evt.data is "DATA2"
[Worker] DONE
PASS successfullyParsed is true

TEST COMPLETE
